MAINT: Cleaned up tests to remove unused dependency injection testing methods.
